Fix get-unique-move-sets test
Essentially, we're giving up on recognizing transitively equivalent
moves because in the general case it's not strictly true that the move
equivalence relation is actually transitive.

It's only transitive when a set of all moves all share mutual
equivalence (like when all 3 combinations of 4 numbers exist as unique
move sets).